feat(price): Replace runtime ToManyStations error with compile time check
This change was done to prevent runtime errors if the user tried to
fetch prices for more than 10 stations at once. Due to a limitation of
the tankerkoenig API it's not possible to request more than 10 stations
at once.

To ensure this behaviour already at compile time, the fetch station
prices function now accepts only an array with up to 10 entries instead
a vector with unknown size.

This comes at costs to the usability, as you know have to wrap the ids
in the array to options, but on the other side it is less likely to
cause runtime errors.